+++ samples/api.callback.js
@@ -1,0 +1,16 @@
+
+const transform = require('..')
+const assert = require('assert')
+
+transform([
+  ['1','2','3','4'],
+  ['a','b','c','d']
+], function(data){
+  data.push(data.shift())
+  return data
+}, function(err, output){
+  assert.deepEqual(output, [
+    [ '2', '3', '4', '1' ],
+    [ 'b', 'c', 'd', 'a' ]
+  ])
+})