+++ samples/module.sync.js
@@ -1,14 +1,14 @@
 
-var transform = require('..');
+const transform = require('..')
 
 transform([
   ['1','2','3','4'],
   ['a','b','c','d']
 ], function(data){
-  data.push(data.shift());
-  return data.join(',')+'\n';
+  data.push(data.shift())
+  return data.join(',')+'\n'
 })
-.pipe(process.stdout);
+.pipe(process.stdout)
 
 // Output:
 // 2,3,4,1